+++ defInt8.go
@@ -37,15 +37,15 @@ func (def *defInt8) value(offset int) (value interface{}, err error) {
 		if def.isNullable {
 			return Int8{IsNull: true}, nil
 		}
-		return int8(0), nil
+		return nil, nil
 	}
 	var int8Value int8
 	on := def.ociNumber[offset]
 	r := C.OCINumberToInt(
 		def.rset.stmt.ses.srv.env.ocierr, //OCIError              *err,
-		&on,                         //const OCINumber       *number,
+		&on,                        //const OCINumber       *number,
 		byteWidth8,                 //uword                 rsl_length,
-		C.OCI_NUMBER_SIGNED,         //uword                 rsl_flag,
+		C.OCI_NUMBER_SIGNED,        //uword                 rsl_flag,
 		unsafe.Pointer(&int8Value)) //void                  *rsl );
 	if r == C.OCI_ERROR {
 		err = def.rset.stmt.ses.srv.env.ociError()