+++ defUint64.go
@@ -37,15 +37,15 @@ func (def *defUint64) value(offset int) (value interface{}, err error) {
 		if def.isNullable {
 			return Uint64{IsNull: true}, nil
 		}
-		return uint64(0), nil
+		return nil, nil
 	}
 	var uint64Value uint64
 	on := def.ociNumber[offset]
 	r := C.OCINumberToInt(
 		def.rset.stmt.ses.srv.env.ocierr, //OCIError              *err,
-		&on,                         //const OCINumber       *number,
-		byteWidth64,                 //uword                 rsl_length,
-		C.OCI_NUMBER_UNSIGNED,         //uword                 rsl_flag,
+		&on,                          //const OCINumber       *number,
+		byteWidth64,                  //uword                 rsl_length,
+		C.OCI_NUMBER_UNSIGNED,        //uword                 rsl_flag,
 		unsafe.Pointer(&uint64Value)) //void                  *rsl );
 	if r == C.OCI_ERROR {
 		err = def.rset.stmt.ses.srv.env.ociError()